Add documentation to Unmarshaller and Marshaller classes
Added comprehensive documentation to the Unmarshaller and Marshaller classes in the src/main/java directory. The documentation includes descriptions of the classes and their methods, as well as detailed explanations of the parameters and return values of the methods. This will aid in understanding the functionality and usage of these classes.