Remove Lombok and Clover dependencies, handle exceptions manually
The Lombok and Clover dependencies have been removed from the project, mainly from the DataFactory.java and the PropertyNearbyConditionClientPerfTest.java files. Exceptions previously handled by Lombok's @SneakyThrows are now handled manually. Other significant changes include updates on the pom.xml file, such as the removal of the openclover and the addition of commons-io in our dependencies.